Add option to share MCache between pipeline runs
MCache must be defined in a separate module instead of being defined in
Compiler.Pipeline.Driver to avoid a cyclic dependency.

(MCache is now used used in Juvix.Compiler.Pipeline (in PipelineEff) and this module is imported by Juvix.Compiler.Pipeline.Driver).
